scenic_runs/requirements.txt
# Core dependencies
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
scanpy>=1.8.0
anndata>=0.8.0
h5py>=3.3.0
tables>=3.6.0

# pySCENIC and related
pyscenic>=0.12.0
arboreto>=0.1.6
ctxcore>=0.2.0
loompy>=3.0.0

# Parallel processing
dask[complete]>=2021.6.0
distributed>=2021.6.0
multiprocessing-logging>=0.3.4

# Utilities
psutil>=5.8.0
memory-profiler>=0.60.0
tqdm>=4.61.0
pyyaml>=5.4.0
adjustText>=0.7.3

# Jupyter
jupyter>=1.0.0
notebook>=6.4.0
ipykernel>=6.0.0
